Service documentation comments and folder restructuring
