+++ src/main/java/com/jaehyune/stamp_app/rest/UserRestController.java
@@ -3,6 +3,7 @@ package com.jaehyune.stamp_app.rest;
 import com.jaehyune.stamp_app.dto.UserCreationDTO;
 import com.jaehyune.stamp_app.dto.UserDTO;
 import com.jaehyune.stamp_app.entity.User;
+import com.jaehyune.stamp_app.rest.error.IdNotFoundException;
 import com.jaehyune.stamp_app.service.UserService;
 import org.springframework.web.bind.annotation.*;
